Fix issues with identifying applicable files
The glob checking code was only checking for the existence of any
matches in the whole file system. So it seemed to work the first time a
file is introduced, but not after that as all existing files were
identifies as exclusions if one file exists that matches the glob.

Various other issues were also fixed:
* Actually skip if the hook is disabled.
* Fix `fileUtils.matchesPath`
* Make a copy of the config object for each hook to avoid dirty
  references
* Set default value of `included` based on presence of a whitelist