Improve performance by running commands simultaneously
With everything running serially and synchronously, affiance is slow.
This was somewhat expected, but in order to get any real usage,
something must be done!
This moves affiance in the direction of allowing hooks to return a
promise that will resolve with the same type of result as before. To
allow for simple hooks that return synchronously, `HookBase.wrapRun`
will also wrap the return value in a promise if necessary.
Finally, since we are now using the `Promise` class, it is time to start
using `es6` features in `affiance`.

* Add `use strict` to the top of files with es6 syntax
* Add `spawnConcurrentCommandsOnApplicableFiles` function that will help
  make shell command based hooks easier to make more performant
* Use `spawnConcurrentCommandsOnApplicableFiles` function in `EsLint`
  hook and `CoffeeLint` hooks for better performance
* Configure `EsLint` to parallelize
* Add `es6` mode to `.eslintrc.yml`