Implement locking infrastructure and update tests
To avoid blowing up the user's machine by spawning too many processes, a
simple, Promise-based locking system has been added to the
`HookContextBase` so that all hooks of a specific type will not use more
than the available slots when spawning commands. Also, the tests for
`CoffeeLint` and `EsLint` were updated to align with new functionality.

* Implement locking infrastructure HookContextBase by exposing
  `waitForProcessSlot` and `releaseProcessSlot` functions.
* Return a promise from `waitForProcessSlot` that either resolves
  immediately with a slotId, or resolved when a slot is released.
* Resolve the next promise in the waiting queue when
  `releaseProcessSlot` is called.
* Update tests for CoffeeLint and EsLint hooks to use promises.