Add new `ignoreMessagePattern` configuration support
Some linters do not provide enough flexibility in their own
configuration files to serve all common use cases. For example, the
`coffeescope2` module of `coffeelint` does not afford configuring
variable naming patterns to ignore with its `unused_var` linter. This is
problematic because coffeescript requires a variable to be defined that
quite often in list comprehension.

This new setting will allow those messages to be suppressed through
`affiance` such that these issues can be worked around without using a
fork of the linter.